+++ src/Information.php
@@ -3,7 +3,7 @@ declare(strict_types = 1);
 
 namespace Innmind\HttpTransport;
 
-use Innmind\Http\Message\{
+use Innmind\Http\{
     Request,
     Response,
 };